requirements.txt
defusedxml>=0.7.1,<0.8.0
google-api-python-client>=2.159.0
python-dotenv>=1.0.1
pytube>=15.0.0
requests>=2.0.0
youtube_transcript_api>=0.6.3